+++ package.json
@@ -28,7 +28,6 @@
     "mathjs": "^12.2.1",
     "nanoid": "^5.0.4",
     "next": "14.0.4",
-    "openai": "^4.52.7",
     "react": "^18",
     "react-dom": "^18",
     "react-draggable": "^4.4.6",
