+++ src/state/canvas.ts
@@ -4,9 +4,6 @@ export type GeneratedCanvas = {
   canvasId: string
   itemId: string
   generatedFromItemId: string
-  lastDrawnAt: number
-  description: string
-  style?: string
 }
 
 export type CanvasStore = {
